feat: 로드맵 생성 관련 타입 수정
